pyproject.toml
[project]
name = "yt-langgraph-shorts"
version = "0.1.0"
description = "Multi-agent YouTube Shorts creator with LangGraph + FastAPI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn[standard]>=0.35.0",
  "langgraph>=0.6.0",
  "langchain-openai>=0.3.0",
  "openai>=1.99.0",
  "python-dotenv>=1.1.1",
  "pydantic>=2.11.0",
  "requests>=2.32.0",
  "moviepy>=1.0.3",
  "Pillow>=11.0.0",
  "gTTS>=2.5.4",
  "loguru>=0.7.2",
  "tqdm>=4.67.0",
]

[tool.uv]
package = false
